+++ source/renderer/app/containers/staking/StakingInfoPage.js
@@ -1,7 +1,6 @@
 // @flow
 import React, { Component } from 'react';
 import { observer, inject } from 'mobx-react';
-import StakingInfo from '../../components/staking/info/StakingInfo';
 import StakingInfoCountdown from '../../components/staking/info/StakingInfoCountdown';
 import type { InjectedProps } from '../../types/injectedPropsType';
 
@@ -14,29 +13,17 @@ export default class StakingInfoPage extends Component<Props> {
 
   render() {
     const { stores, actions } = this.props;
-    const {
-      decentralizationProgress,
-      epochToFullyDecentralized,
-      isFullyDecentralized,
-    } = stores.networkStatus;
+    const { isAlonzoActivated, alonzoEraEpoch } = stores.networkStatus;
     const { stakingInfoWasOpen } = stores.staking;
     const { setStakingInfoWasOpen } = actions.staking;
     const { openExternalLink } = stores.app;
-    if (!epochToFullyDecentralized) {
-      return (
-        <StakingInfo
-          percentage={decentralizationProgress}
-          onLearnMoreClick={openExternalLink}
-        />
-      );
-    }
     return (
       <StakingInfoCountdown
         percentage={decentralizationProgress}
         onLearnMoreClick={openExternalLink}
-        epoch={epochToFullyDecentralized}
+        epoch={alonzoEraEpoch}
         onSetStakingInfoWasOpen={setStakingInfoWasOpen.trigger}
-        isFullyDecentralized={isFullyDecentralized}
+        isAlonzoActivated={isAlonzoActivated}
         stakingInfoWasOpen={stakingInfoWasOpen}
       />
     );