+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Changelog
 
 ### Chores
 
+- Updated "Decentralization Info" screen text copy and info for the upcoming Alonzo upgrade ([PR ](TBD))
 - Updated `cardano-wallet` to version `2021-07-30` ([PR 2635](https://github.com/input-output-hk/daedalus/pull/2635), [PR 2638](https://github.com/input-output-hk/daedalus/pull/2638))
 - Removed "Decentralization countdown" tada icon and info tab ([PR 2625](https://github.com/input-output-hk/daedalus/pull/2625))